\name{packageLab03-package}
\alias{packageLab03-package}
\alias{packageLab03}
\docType{package}
\title{
\packageTitle{packageLab03}
}
\description{
\packageDescription{packageLab03}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{packageLab03}
\packageIndices{packageLab03}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{packageLab03}

Maintainer: \packageMaintainer{packageLab03}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
